requirements.txt
opencv-python
mediapipe
scikit-learn
streamlit
numpy 
Pillow